requirements.txt
django>=1.8
django-cms>=3.0
